fix: more adjustments, fix build
